+++ library/ConsoleLib.py
@@ -10,6 +10,7 @@ class Time:
 
     Created by: Raymond Allison
     """
+    
     def timeFormat():
         """
         Sets the time format and returns a String version for the console.
@@ -23,7 +24,6 @@ class Time:
         date_object = datetime.now()
         formTime = date_object.strftime("%H:%M:%S - %b %d %Y")
         return formTime
-
     def timeFormatUniversial():
         """
         Sets the time format, UTC.
@@ -69,12 +69,10 @@ class Essentials:
         ]
         response8ball = random.choice(responses)
         return response8ball
-
     def ballImage():
         """Just returns an 8ball image."""
         ballImage = 'https://cdn.discordapp.com/attachments/576096750331494420/896414018133196800/8b.png'
         return ballImage
-
     def CoinToss():
         """
         Responses handler for a randomized coinflip.