+++ library/MemeSelect.py
@@ -1,0 +1,15 @@
+import random 
+
+class MemeList:
+    def Renault():
+        """Originally a meme."""
+        renault = 'https://cdn.discordapp.com/attachments/879488846453174282/895849572033781822/renault.png'
+        return renault
+    def Biden():
+        """Load of Malarkey."""
+        biden = 'https://cdn.discordapp.com/attachments/576096750331494420/897532315666898994/P20210303AS-1901-cropped.png'
+        return biden
+    def Trump():
+        """I'm really rich, like really rich."""
+        trump = 'https://cdn.discordapp.com/attachments/576096750331494420/898227351702343700/abfb424b-43ab-40a7-b360-ca5750941f72-AP21251549466117.png'
+        return trump